Move functionality to registry-sync-v0 and apply copilot's suggestions
