--- fabric-modlist-info-v1/build.gradle
@@ -1,14 +1,0 @@
-version = getSubprojectVersion(project)
-
-loom {
-	accessWidenerPath = file("src/main/resources/fabric-modlist-info-v1.classtweaker")
-}
-
-moduleDependencies(project, [
-	'fabric-api-base',
-	'fabric-lifecycle-events-v1'
-])
-
-testDependencies(project, [
-	':fabric-lifecycle-events-v1'
-])
