--- fabric-modlist-info-v1/src/main/java/net/fabricmc/fabric/impl/modlist/ModListSaveInit.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2016, 2017, 2018, 2019 FabricMC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package net.fabricmc.fabric.impl.modlist;
-
-import net.fabricmc.api.ModInitializer;
-import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
-
-/**
- * Entrypoint for the Fabric Mod List Info module.
- *
- * <p>Registers a {@link ServerLifecycleEvents#AFTER_SAVE} listener that writes
- * the active mod list to the world's save directory AFTER every save.
- *
- * @see ModListSaver
- */
-public class ModListSaveInit implements ModInitializer {
-	@Override
-	public void onInitialize() {
-		ServerLifecycleEvents.AFTER_SAVE.register((server, flush, force) -> ModListSaver.save(server));
-	}
-}
