--- fabric-modlist-info-v1/src/main/resources/fabric-modlist-info-v1.classtweaker
@@ -1,2 +1,0 @@
-classTweaker	v1	official
-accessible	field	net/minecraft/server/MinecraftServer	storageSource	Lnet/minecraft/world/level/storage/LevelStorageSource$LevelStorageAccess;
