--- fabric-modlist-info-v1/src/main/resources/fabric.mod.json
@@ -1,33 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-modlist-info-v1",
-  "name": "Fabric Mod List Info (v1)",
-  "version": "${version}",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "icon": "assets/fabric-modlist-info-v1/icon.png",
-  "contact": {
-    "homepage": "https://fabricmc.net",
-    "irc": "irc://irc.esper.net:6667/fabric",
-    "issues": "https://github.com/FabricMC/fabric/issues",
-    "sources": "https://github.com/FabricMC/fabric"
-  },
-  "authors": [
-    "FabricMC"
-  ],
-  "depends": {
-    "fabricloader": ">=0.18.4",
-    "fabric-api-base": "*",
-    "fabric-lifecycle-events-v1": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.impl.modlist.ModListSaveInit"
-    ]
-  },
-  "description": "Saves a list of active mods into the world save directory.",
-  "custom": {
-    "fabric-api:module-lifecycle": "experimental"
-  },
-  "accessWidener": "fabric-modlist-info-v1.classtweaker"
-}
