--- fabric-modlist-info-v1/src/testmod/resources/fabric.mod.json
@@ -1,16 +1,0 @@
-{
-  "schemaVersion": 1,
-  "id": "fabric-modlist-info-v1-testmod",
-  "name": "Fabric Mod List Info (v1) Test Mod",
-  "version": "1.0.0",
-  "environment": "*",
-  "license": "Apache-2.0",
-  "depends": {
-    "fabric-modlist-info-v1": "*"
-  },
-  "entrypoints": {
-    "main": [
-      "net.fabricmc.fabric.test.modlist.ModListInfoTest"
-    ]
-  }
-}
