+++ fabric-registry-sync-v0/build.gradle
@@ -6,7 +6,8 @@ loom {
 
 moduleDependencies(project, [
 	'fabric-api-base',
-	'fabric-networking-api-v1'
+	'fabric-networking-api-v1',
+	'fabric-lifecycle-events-v1'
 ])
 
 testDependencies(project, [
