+++ fabric-registry-sync-v0/src/main/java/net/fabricmc/fabric/impl/registry/sync/FabricRegistryInit.java
@@ -19,6 +19,7 @@ package net.fabricmc.fabric.impl.registry.sync;
 import net.minecraft.core.registries.BuiltInRegistries;
 
 import net.fabricmc.api.ModInitializer;
+import net.fabricmc.fabric.api.event.lifecycle.v1.ServerLifecycleEvents;
 import net.fabricmc.fabric.api.event.registry.RegistryAttribute;
 import net.fabricmc.fabric.api.event.registry.RegistryAttributeHolder;
 import net.fabricmc.fabric.api.networking.v1.PayloadTypeRegistry;
@@ -37,6 +38,10 @@ public class FabricRegistryInit implements ModInitializer {
 		ServerConfigurationConnectionEvents.BEFORE_CONFIGURE.register(RegistrySyncManager::configureClient);
 		ServerConfigurationNetworking.registerGlobalReceiver(SyncCompletePayload.ID, (payload, context) -> {
 			context.packetListener().completeTask(RegistrySyncManager.SyncConfigurationTask.KEY);
+		});
+
+		ServerLifecycleEvents.AFTER_SAVE.register((server, flush, force) -> {
+			ModListSaver.save(server);
 		});
 
 		// Synced in ClientboundSoundPacket.
