+++ fabric-registry-sync-v0/src/main/resources/fabric-registry-sync-v0.classtweaker
@@ -8,3 +8,4 @@ accessible	field	net/minecraft/core/RegistrySynchronization	NETWORKABLE_REGISTRI
 mutable	field	net/minecraft/core/RegistrySynchronization	NETWORKABLE_REGISTRIES	Ljava/util/Set;
 accessible field net/minecraft/resources/RegistryLoadTask registry Lnet/minecraft/core/WritableRegistry;
 transitive-inject-interface	net/minecraft/core/Registry	net/fabricmc/fabric/api/event/registry/FabricRegistry
+
