+++ fabric-registry-sync-v0/src/testmod/resources/fabric.mod.json
@@ -15,7 +15,8 @@
     "main": [
       "net.fabricmc.fabric.test.registry.sync.CustomDynamicRegistryTest",
       "net.fabricmc.fabric.test.registry.sync.RegistrySyncTest",
-      "net.fabricmc.fabric.test.registry.sync.RegistryAliasTest"
+      "net.fabricmc.fabric.test.registry.sync.RegistryAliasTest",
+      "net.fabricmc.fabric.test.registry.sync.ModListInfoTest"
     ],
     "client": [
       "net.fabricmc.fabric.test.registry.sync.client.DynamicRegistryClientTest",
