+++ gradle.properties
@@ -57,4 +57,3 @@ fabric-tag-api-v1-version=2.1.3
 fabric-transfer-api-v1-version=8.0.11
 fabric-transitive-access-wideners-v1-version=8.1.4
 fabric-convention-tags-v2-version=4.7.0
-fabric-modlist-info-v1-version=1.0.0
