+++ settings.gradle
@@ -50,7 +50,6 @@ include 'fabric-loot-api-v3'
 include 'fabric-menu-api-v1'
 include 'fabric-message-api-v1'
 include 'fabric-model-loading-api-v1'
-include 'fabric-modlist-info-v1'
 include 'fabric-networking-api-v1'
 include 'fabric-object-builder-api-v1'
 include 'fabric-particles-v1'
